Moved alias and tag support to abstract container class
