+++ src/Container.php
@@ -141,70 +141,6 @@ class Container implements \ArrayAccess, ContainerInterface, ResetInterface
     }
 
     /**
-     * Marks an alias id to service id.
-     *
-     * @param string $id The alias id
-     * @param string $serviceId The registered service id
-     *
-     * @throws ContainerResolutionException Service id is not found in container
-     */
-    public function alias(string $id, string $serviceId): void
-    {
-        if ($id === $serviceId) {
-            throw new \LogicException("[{$id}] is aliased to itself.");
-        }
-
-        if (!$this->offsetExists($serviceId)) {
-            throw new ContainerResolutionException("Service id '{$serviceId}' is not found in container");
-        }
-
-        $this->aliases[$id] = $this->aliases[$serviceId] ?? $serviceId;
-    }
-
-    /**
-     * Assign a set of tags to service(s).
-     *
-     * @param string[]|string         $serviceIds
-     * @param array<int|string,mixed> $tags
-     */
-    public function tag($serviceIds, array $tags): void
-    {
-        foreach ((array) $serviceIds as $service) {
-            foreach ($tags as $tag => $attributes) {
-                // Exchange values if $tag is an integer
-                if (\is_int($tmp = $tag)) {
-                    $tag = $attributes;
-                    $attributes = $tmp;
-                }
-
-                $this->tags[$service][$tag] = $attributes;
-            }
-        }
-    }
-
-    /**
-     * Resolve all of the bindings for a given tag.
-     *
-     * @param string $tag
-     *
-     * @return mixed[] of [service, attributes]
-     */
-    public function tagged(string $tag): array
-    {
-        $tags = [];
-
-        foreach ($this->tags as $service => $tagged) {
-            if (isset($tagged[$tag])) {
-                $tags[] = [$this->get($service), $tagged[$tag]];
-            }
-        }
-
-        return $tags;
-    }
-
-    /**
-     * Wraps a class string or callable with a `call` method.
-     *
      * This is useful when you want to autowire a callable or class string lazily.
      *
      * @param callable|string $definition A class string or a callable
